We need to make sure we pick up the right versions of casperjs and phantomjs.
Before this, we were failing with the error message "casperjs needs
phantomjs 1.x.x to run".

I could have solved this problem by upgrading casperjs as well, but I
was having trouble getting npm to do what I wanted it to do, and in
any case this solution is better since it keeps the (existing working)
behavior the same.

Resolves https://app.asana.com/0/27216215224639/73642614666076/f

Auditors: eater

Test plan:
Ran
   ./exercise-icons/full-run.sh
on toby with this change, and saw it say:
   Screenshotting [-------------------------------------------------] 0% 0.0 1/4173
whereas before it would just give an error.